backend/requirements.txt
# Document processing
docling>=2.55.0
langchain-docling>=0.1.0

# LangChain / LangGraph
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langchain-chroma>=0.1.0

# Vector store
chromadb>=0.4.22

# FastAPI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.0.0

# Config & utilities
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy<2